dev-build | refactor code.

+++ Identity.Application/Services/UserService.cs
@@ -1,12 +1,15 @@
 ﻿using AutoMapper;
 using Identity.Application.Configurations.Settings;
+using Identity.Application.Dtos;
 using Identity.Application.Dtos.Users;
 using Identity.Application.Interfaces.Repositories;
 using Identity.Application.Interfaces.Services;
+using Identity.Domain.Common;
 using Identity.Domain.Entities;
 using Identity.Domain.Exceptions;
 using Identity.Domain.Helpers;
 using Mapster;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Options;
@@ -20,17 +23,25 @@ namespace Identity.Application.Services
             ITokenRepository tokenRepository,
             IRefreshTokenRepository refreshTokenRepository,
             IOptions<ApplicationSettings> applicationSettings,
-            IMapper mapper) : base(userManager, passwordHasher, refreshTokenRepository, tokenRepository, applicationSettings, mapper)
+            IMapper mapper,
+            IHttpContextAccessor httpContextAccessor) : base(userManager, passwordHasher, refreshTokenRepository, tokenRepository, applicationSettings, mapper, httpContextAccessor)
         {
         }
 
-        public async Task<bool> DeleteByIdAsync(string id)
+        public async Task<bool> DeleteByIdAsync(string id, bool isSoftDelete = true)
         {
             User? currentUser = await _userManager.FindByIdAsync(id) ?? throw new NotFoundException($"User with {id} not found!");
 
-            IdentityResult identityResult = await _userManager.DeleteAsync(currentUser);
+            if (!isSoftDelete)
+            {
+                IdentityResult identityResult = await _userManager.DeleteAsync(currentUser);
 
-            return identityResult.Succeeded;
+                return identityResult.Succeeded;
+            }
+
+            currentUser.IsDeleted = true;
+
+            return (await _userManager.UpdateAsync(currentUser)).Succeeded;
         }
 
         public async Task<IList<UserDto>> GetAllAsync()
@@ -98,6 +109,25 @@ namespace Identity.Application.Services
             return true;
         }
 
+        public async Task<PaginatedResponse<UserDto>> SearchAsync(SearchRequest request, bool isIncludeDeletedUser = false)
+        {
+            string keyword = request.Keyword;
+
+            IQueryable<User> query = _userManager.Users.AsQueryable();
+
+            Func<IQueryable<User>, IQueryable<User>>? predicate = null;
+            if (!string.IsNullOrEmpty(keyword))
+            {
+                predicate = (user) => user.Where(x => x.DisplayName.Contains(keyword)
+                                                   || x.FirstName.Contains(keyword)
+                                                   || x.LastName.Contains(keyword));
+
+                query = predicate(query);
+            }
+
+            return await PaginatedResponse<UserDto>.CreateAsync(query.Adapt<IQueryable<UserDto>>(), request.PageIndex, request.PageSize);
+        }
+
         public async Task<bool> SendResetPasswordEmailAsync(string email)
         {
             User? user = await _userManager.FindByEmailAsync(email) ?? throw new ArgumentException($"User with {email} doesn't exist.");
@@ -118,7 +148,10 @@ namespace Identity.Application.Services
 
             currentUser = request.Adapt(currentUser);
 
-            await ValidateModelAsync(currentUser, request.Password);
+            await ValidateModelAsync(requestModel: currentUser, password: string.Empty, isUpdate: true);
+
+            currentUser.ModifiedBy = !string.IsNullOrEmpty(LoginSession?.Email) ? LoginSession.Email : currentUser.ModifiedBy;
+            currentUser.ModifiedDate = DateTime.UtcNow;
 
             IdentityResult? updateResult = await _userManager.UpdateAsync(currentUser);
 
@@ -137,6 +170,8 @@ namespace Identity.Application.Services
             ValidatePassword(newPassword);
 
             currentUser.PasswordHash = _passwordHasher.HashPassword(currentUser, newPassword);
+            currentUser.ModifiedBy = !string.IsNullOrEmpty(LoginSession?.Email) ? LoginSession.Email : currentUser.ModifiedBy;
+            currentUser.ModifiedDate = DateTime.UtcNow;
 
             IdentityResult? identityResult = await _userManager.UpdateAsync(currentUser);
